[WSW-65549]Fix compile err of ncp-device example
[Description]
Fix compile error of ncp-device of usb interface.
[Fix]
Fix compile error of ncp-device of usb interface.
[Testing]
ncp device with usb interface compile pass.
[Reviewers]
Gannicus Guo, Gang Li

+++ ncp_examples/common/ncp_adapter/ncp_intf/usb/usb_host_cdc/ncp_intf_usb_host_cdc.c
@@ -200,13 +200,13 @@ int ncp_usb_host_deinit(void* argv)
     return ret;
 }
 
-static int ncp_usb_host_pm_enter(int32_t pm_state)
+static int ncp_usb_host_pm_enter(uint8_t pm_state)
 {
     /* TODO: NCP usb pm */
     return NCP_STATUS_SUCCESS;
 }
 
-static int ncp_usb_host_pm_exit(int32_t pm_state)
+static int ncp_usb_host_pm_exit(uint8_t pm_state)
 {
     /* TODO: NCP usb pm */
     return NCP_STATUS_SUCCESS;
